+++ modules/canvaskit/CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
  - `SkCanvas.concat44` has been folded into concat (which now takes 3x2, 3x3, or 4x4 matrices). It will
    be removed soon.
+   
+### Fixed
+ - Memory leak in paragraph binding code (https://github.com/flutter/flutter/issues/56938)
 
 ## [0.14.0] - 2020-03-18